requirements.txt
torch
torchvision
monai
monai_generative
numpy
matplotlib
tqdm
wandb
Pillow
lpips
git+https://github.com/MIC-DKFZ/MedNeXt.git